% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{autoplot.hagis.complexities}
\alias{autoplot.hagis.complexities}
\title{Plot hagis Complexities Object}
\usage{
\method{autoplot}{hagis.complexities}(object, type, color = NULL,
  order = NULL, ...)
}
\arguments{
\item{object}{a \pkg{hagis} \code{complexities} object generated by
\code{\link[=calculate_complexities]{calculate_complexities()}}. Character.}

\item{type}{a vector of values for which the bar plot is desired. Specify
whether to return a graph of the frequency of complexities as a percentage,
"\code{percentage}", or as the count, "\code{count}". Character.}

\item{color}{a named or hexadecimal color value to use for the bar color}

\item{order}{sort the x-axis of the bar chart by ascending or descending
order of \code{frequency}. Accepts \code{ascending} or \code{descending} input values.
Defaults to \code{complexity} value. Character.}

\item{...}{passed to the chosen \code{geom(s)}}
}
\value{
A \link[ggplot2]{ggplot2} plot
}
\description{
Creates a \pkg{ggplot2} object of the frequency of complexity
(percent per complexity) or a \pkg{ggplot2} object of the distribution
(number per complexity) calculated by \code{\link[=calculate_complexities]{calculate_complexities()}}.
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calculate_complexities(x = Ps,
                                       cutoff = 60,
                                       control = "susceptible",
                                       sample = "Isolate",
                                       gene = "Rps",
                                       perc_susc = "perc.susc")

# Visualize the distribution (count or actual values)
autoplot(x = complexities, type = "count")

# Visualize the frequency (percentages)
autoplot(x = complexities, type = "percentage")

}
