% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{calc_complexities}
\alias{calc_complexities}
\title{Calculate Distribution of Complexities by Pathogenicity Gene}
\usage{
calc_complexities(x, cutoff, control = "susceptible",
  sample = "sample", gene = "gene")
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Numeric.}

\item{control}{Value used to denote the susceptible control in the \code{gene}
field. Character.}

\item{sample}{Field providing the unique isolate identification for each
isolate being tested. Character.}

\item{gene}{Field providing the pathogenicty gene(s) being tested. Character.}
}
\description{
This function will calculate the distribution of
susceptibilities by pathogenicity gene.
}
\examples{


# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calc_complexities(Ps, cutoff = 60)
complexities

}
