% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{calculate_complexities}
\alias{calculate_complexities}
\title{Calculate Distribution of Complexities by Sample}
\usage{
calculate_complexities(x, cutoff, control, sample, gene, perc_susc)
}
\arguments{
\item{x}{a \code{data.frame} containing the data.}

\item{cutoff}{value for percent susceptible cutoff. Numeric.}

\item{control}{value used to denote the susceptible control in the \code{gene}
field. Character.}

\item{sample}{field providing the unique identification for each sample being
tested. Character.}

\item{gene}{field providing the gene(s) being tested. Character.}

\item{perc_susc}{field providing the percent susceptible reactions.
Character.}
}
\value{
\code{calculate_complexities} returns an object of class
\code{hagis.complexities}.

An object of class \code{hagis.complexities} is a \code{list} containing the following
components
\describe{
\item{grouped_complexities}{a \code{\link[data.table:data.table]{data.table::data.table()}} object of
grouped complexities}
\item{individual_complexities}{a \code{\link[data.table:data.table]{data.table::data.table()}} object of
individual complexities}
}
}
\description{
This function will calculate the distribution of
susceptibilities by sample
}
\examples{

# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calculate_complexities(x = Ps,
                                      cutoff = 60,
                                      control = "susceptible",
                                      sample = "Isolate",
                                      gene = "Rps",
                                      perc_susc = "perc.susc")
complexities

summary(complexities)

# plot complexities frequency or by percentage
plot(complexities, type = "percentage")

}
