% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_index.R
\name{calculate_diversity}
\alias{calculate_diversity}
\title{Calculate Diversity Index}
\usage{
calculate_diversity(x, cutoff, control, sample, Rps, perc_susc)
}
\description{
Calculates \emph{Phytophthora} diversity index
}
