% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{complexities}
\alias{complexities}
\title{Calculate distribution of complexities by Rps gene}
\usage{
complexities(x, sample, percent.susc, gene, susceptibility_cutoff)
}
\arguments{
\item{x}{A \code{data.frame} containing the data}

\item{susceptibility_cutoff}{}
}
\description{
This function will calculate the distribution of
susceptibilities by Rps gene.
}
