% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_index.R
\name{diversity}
\alias{diversity}
\title{Calculate Diversity Index}
\usage{
diversity(x, cutoff, control = "susceptible")
}
\arguments{
\item{x}{A \code{data.frame} containing the data}

\item{cutoff}{Value for percent susceptible cutoff. Integer.}

\item{control}{Value used to denote the susceptible control in the \emph{Rps}
column. Defaults to "susceptible". Character.}
}
\description{
Calculates \emph{Phytophthora} diversity index
}
