% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{diversity_table.hagis.diversities}
\alias{diversity_table.hagis.diversities}
\title{Prints  hagis.diversities Object Table of Diversities}
\usage{
\method{diversity_table}{hagis.diversities}(x, ...)
}
\arguments{
\item{x}{a hagis.diversities object}

\item{...}{ignored}
}
\description{
Custom S3 method for printing the frequency table of diversities from a
\code{hagis.diversities} object
}
\examples{
# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
diversities <- calculate_diversities(x = Ps,
                                     cutoff = 60,
                                     control = "susceptible",
                                     sample = "Isolate",
                                     Rps = "Rps",
                                     perc_susc = "perc.susc")

#' # print the diversities table
diversity_table(diversities)
}
