% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathotype_frequency_distribution.R
\name{freq_dist}
\alias{freq_dist}
\title{Calculate pathotype frequency distribution}
\usage{
freq_dist(x, cutoff)
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Integer.}
}
\description{
Calculates \emph{Phytophthora} pathotype frequency distribution
}
