% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hagis-package.R
\docType{package}
\name{hagis-package}
\alias{hagis}
\alias{hagis-package}
\title{hagis: Analysis of Plant Pathogen Pathotype Complexities, Distributions and Diversity}
\description{
Analysis of plant pathogen pathotype survey data. Functions provided calculate distribution of susceptibilities, distribution of complexities with statistics, pathotype frequency distribution, as well as diversity indices for pathotypes. This package is meant to be a direct replacement for Herrmann, Löwer and Schachtel's (1999) <doi:10.1046/j.1365-3059.1999.00325.x> Habgood-Gilmour Spreadsheet, 'HaGiS', previously used for pathotype analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openplantpathology/hagis}
  \item \url{https://openplantpathology.github.io/hagis/}
  \item Report bugs at \url{https://github.com/openplantpathology/hagis/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Austin G. McCoy \email{mccoyaus@msu.edu} (\href{https://orcid.org/0000-0003-2483-4184}{ORCID}) [conceptor]
  \item Zachary Noel \email{noelzach@msu.edu} (\href{https://orcid.org/0000-0001-6375-8300}{ORCID}) [conceptor]
  \item Martin Chilvers \email{chilvers@msu.edu } (\href{https://orcid.org/0000-0001-8832-1666}{ORCID})
}

Other contributors:
\itemize{
  \item Zhian N. Kamvar (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [contributor, reviewer]
  \item Michigan Soybean Promotion Committee [funder]
  \item Project GREEEN [funder]
  \item North Central Soybean Research Program [funder]
  \item GRDC Project DAQ00186 [funder]
}

}
\keyword{internal}
