% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hagis-package.R
\docType{package}
\name{hagis}
\alias{hagis}
\alias{hagis-package}
\title{Tools for analysis of plant pathogen pathotype complexities, distributions and diversity}
\description{
Analysis of plant pathogen pathotype survey data.  Functions
  provided calculate distribution of susceptibilities, distribution of
  complexities with statistics, pathotype frequency distribution, as well as
  diversity indices for pathotypes.  This package is meant to be a direct
  replacement for Herrmann, Löwer, Schachtel's (1999)
  <doi:10.1046/j.1365-3059.1999.00325.x> Habgood-Gilmour Spreadsheet, 'HaGiS',
  previously used for pathotype analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openplantpathology/hagis}
  \item \url{https://openplantpathology.github.io/hagis/}
  \item Report bugs at \url{https://github.com/openplantpathology/hagis/issues}
}

}
\author{
Austin G. McCoy, mccoyaus@msu.edu and Zachary Noel, noelzach@msu.edu and
Adam H. Sparks, adam.sparks@usq.edu.au and Martin Chilvers, chilvers@msu.edu
}
