% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{individual_pathotypes}
\alias{individual_pathotypes}
\title{Prints Individual Pathotypes for Each Sample}
\usage{
individual_pathotypes(x, type = "text", ...)
}
\arguments{
\item{x}{a hagis.diversities object generated by \code{\link[=calculate_diversities]{calculate_diversities()}}}

\item{type}{return a text or graphic type table. Valid values are \code{text} or
\code{graphic} Defaults to \code{text}.}

\item{...}{other arguments passed to \code{\link[pander:panderOptions]{pander::panderOptions()}} or
\code{\link[gridExtra:ttheme_default]{gridExtra::ttheme_default()}}}
}
\description{
Print an object from a \code{hagis.diversities} object with individual pathotypes,
\emph{i.e.} each sample's pathotype.
Users may select from a \link{pander} table (a text object for Markdown) or a
graphic object (a \code{\link[gridExtra:tableGrob]{gridExtra::tableGrob()}}).
}
\examples{
# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
diversities <- calculate_diversities(x = Ps,
                                     cutoff = 60,
                                     control = "susceptible",
                                     sample = "Isolate",
                                     Rps = "Rps",
                                     perc_susc = "perc.susc")

# print the diversities table
individual_pathotypes(diversities)

}
\seealso{
\code{\link[=calculate_diversities]{calculate_diversities()}}, \code{\link[=diversities_table]{diversities_table()}}
}
