% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_sojae_survey}
\alias{P_sojae_survey}
\title{\emph{Phytophthora sojae} Survey Example Data}
\format{An object of class \code{data.table} with 294 observations of 12 variables
\describe{
\item{Isolate}{\emph{P. sojae} isolate identifier}
\item{Line}{Soybean cultivar}
\item{Rps}{\emph{Rps} gene identifer}
\item{Total}{}
\item{HR (1)}{}
\item{Lesion (2)}{}
\item{Lesion to cotyledon (3)}{}
\item{Dead (4)}{}
\item{total.susc}{}
\item{total.resis}{}
}}
\source{
Austin McCoy's experiment to be supplied by Austin here...
}
\usage{
data(P_sojae_survey)
}
\description{
Data from a \emph{Phytophthora sojae} survey
}
\examples{
data(P_sojae_survey)
P_sojae_survey

}
\keyword{datasets}
