% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents.R
\name{percents}
\alias{percents}
\title{Calculate Total and Percent Resistant and Susceptible Reactions}
\usage{
percents(x)
}
\arguments{
\item{x}{A \code{data.frame} containing the data. See Details below for more.
Character.}
}
\value{
A \code{data.frame} as a \code{data.table} object with total and percent
susceptible and resistant values
}
\description{
This function will calculate total percent resistant and
susceptible reactions for each \emph{Rps} gene
}
\details{
The \code{data.frame} object supplied must contain fields with the following
names and contents where each line is a unique observation.
\itemize{
\item isolate - Unique identifier for the isolate
\item Rps - \emph{Rps} gene(s) included in study
\item total - Total number of plants
\item hr - Number of plants with hypersensitive response to inoculation
\item lesion - Number of plants with lesion in response to inoculation
\item lesion_to_cotyledon - Number of plants with lesion to cotyledon
in response to inoculation
\item dead - Number of plants dead in response to inoculation
}

The use of \code{\link[=write_template]{write_template()}} is encouraged to ensure that the input file
used is properly formatted.
}
\examples{

# locate system file for import
x <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
y <- read.csv(x)

# calculate totals and percents for resistant and susceptible reactions
z <- percents(y)
z

}
\author{
Adam H. Sparks, adamhsparks@gmail.com
}
