% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rps.R
\name{plot.hagis.rps.summary}
\alias{plot.hagis.rps.summary}
\title{Plot \emph{Rps} Gene Summaries}
\usage{
\method{plot}{hagis.rps.summary}(data)
}
\arguments{
\item{data}{a a \pkg{hagis} \code{rps.summary} object generated by
\code{\link[=summarize_rps]{summarize_rps()}}. Character.}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Creates a \pkg{ggplot2} object of the \emph{Rps} gene
summaries calculated by \code{\link[=summarize_rps]{summarize_rps()}}
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_rps(x = Ps,
                       cutoff = 60,
                       control = "susceptible",
                       sample = "Isolate",
                       Rps = "Rps",
                       perc_susc = "perc.susc")

# Visualize the summary of Rps genes
plot(susc)
}
\seealso{
\code{\link[=summarize_rps]{summarize_rps()}}
}
