% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_meta}
\alias{sample_meta}
\title{\emph{Phytophthora sojae} Survey Example Metadata}
\format{
An object of class \code{data.frame} with 21 observations of 2 variables
\describe{
\item{Sample}{\emph{P. sojae} sample identifier}
\item{Locale}{US State where sample was collected, limited to Michigan}
}
}
\source{
Data from an ongoing 2017 \emph{Phytophthora sojae} survey in Michigan,
conducted by A. G. McCoy \emph{et al.}.
}
\usage{
data(sample_meta)
}
\description{
Metadata to accompany internal data from a \emph{Phytophthora sojae} survey used
for an example in the "Beta-diversity Analyses" vignette.
}
\examples{
data(sample_meta)
sample_meta

}
\keyword{datasets}
