% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rps.R
\name{summarize_rps}
\alias{summarize_rps}
\title{Calculate and Summarize Distribution of Susceptibilities by \emph{Rps} Gene}
\usage{
summarize_rps(x, cutoff, control, sample, Rps, perc_susc)
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Numeric.}

\item{control}{Value used to denote the susceptible control in the \code{Rps}
field. Character.}

\item{sample}{Field providing the unique isolate identification for each
isolate being tested. Character.}

\item{Rps}{Field providing the \emph{Rps} gene(s) being tested. Character.}

\item{perc_susc}{Field providing the percent susceptible reactions.
Character.}
}
\description{
This function will calculate the distribution of
susceptibilities by \emph{Rps} gene.
}
\examples{

# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_rps(x = Ps,
                       cutoff = 60,
                       control = "susceptible",
                       sample = "Isolate",
                       Rps = "Rps",
                       perc_susc = "perc.susc")
susc

}
\seealso{
\code{\link[=write_template]{write_template()}}, \code{\link[=percents]{percents()}}, \code{\link[=visualize_rps]{visualize_rps()}}
}
