% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{summary.hagis}
\alias{summary.hagis}
\title{Summarises \pkg{hagis} Complexity Object}
\usage{
\method{summary}{hagis}(object, ...)
}
\arguments{
\item{...}{ignored}

\item{x}{A hagis class object (a list of two \code{data.table}s)}
}
\description{
Custom \code{\link[=summary]{summary()}} method for \pkg{hagis} \code{complexity} objects.
}
