% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{visualize_complexities}
\alias{visualize_complexities}
\title{Visualize Complexities}
\usage{
visualize_complexities(x)
}
\arguments{
\item{x}{A \code{data.frame} generated by \code{\link[=calculate_complexities]{calculate_complexities()}}. Character.}
}
\description{
Creates a \pkg{ggplot2} object of the frequency of complexity
(% per complexity) calculated by \code{\link[=calculate_complexities]{calculate_complexities()}}
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- calculate_complexities(x = Ps,
                          cutoff = 60,
                          control = "susceptible",
                          sample = "Isolate",
                          Rps = "Rps",
                          perc_susc = "perc.susc")

# Visualize the distribution (actual values)
visualize_complexities(susc)

}
\seealso{
\code{\link[=calculate_complexities]{calculate_complexities()}}, \code{\link[=visualize_distribution]{visualize_distribution()}}
}
