% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rps.R
\name{visualize_rps}
\alias{visualize_rps}
\title{Visualize \emph{Rps} Gene Summaries}
\usage{
visualize_rps(x)
}
\arguments{
\item{x}{A \code{data.frame} generated by \code{\link[=summarize_rps]{summarize_rps()}}. Character.}
}
\description{
Creates a \pkg{ggplot2} object of the \emph{Rps} gene
summaries calculated by \code{\link[=summarize_rps]{summarize_rps()}}
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_rps(x = Ps,
                       cutoff = 60,
                       control = "susceptible",
                       sample = "Isolate",
                       Rps = "Rps",
                       perc_susc = "perc.susc")

# Visualize the summary of Rps genes
visualize_rps(susc)

}
\seealso{
\code{\link[=write_template]{write_template()}}, \code{\link[=calc_percents]{calc_percents()}}, \code{\link[=summarize_rps]{summarize_rps()}}
}
