% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{calculate_complexities}
\alias{calculate_complexities}
\title{Calculate Distribution of Complexities by Isolate}
\usage{
calculate_complexities(x, cutoff, control, sample, Rps, perc_susc)
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Numeric.}

\item{control}{Value used to denote the susceptible control in the \code{Rps}
field. Character.}

\item{sample}{Field providing the unique isolate identification for each
isolate being tested. Character.}

\item{Rps}{Field providing the \emph{Rps} gene(s) being tested. Character.}

\item{perc_susc}{Field providing the percent susceptible reactions.
Character.}
}
\value{
\code{calculate_complexities} returns an object of \code{\link[=class]{class()}} "hagis"
An object of class "hagis" is a list containing the following components
\itemize{
\item{grouped_complexities}{a \link[=data.table]{data.table} object of
grouped complexities}
\item{individual_complexities}{a \link[=data.table]{data.table} object of
individual complexities}
}
}
\description{
This function will calculate the distribution of
susceptibilities by isolate.
}
\examples{

# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calculate_complexities(x = Ps,
                       cutoff = 60,
                       control = "susceptible",
                       sample = "Isolate",
                       Rps,
                       perc_susc = "perc.susc")
summary(complexities)

}
\seealso{
\code{\link[=visualize_distribution]{visualize_distribution()}}, \code{\link[=visualize_complexity]{visualize_complexity()}}
}
