% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{calculate_diversities}
\alias{calculate_diversities}
\title{Calculate diversities indices}
\usage{
calculate_diversities(x, cutoff, control, sample, gene, perc_susc)
}
\arguments{
\item{x}{a \code{data.frame} containing the data.}

\item{cutoff}{value for percent susceptible cutoff. Numeric.}

\item{control}{value used to denote the susceptible control in the \code{gene}
field. Character.}

\item{sample}{field providing the unique identification for each sample being
tested. Character.}

\item{gene}{field providing the gene(s) being tested. Character.}

\item{perc_susc}{field providing the percent susceptible reactions.
Character.}
}
\value{
hagis.diversities object containing
\itemize{
\item Number of Samples
\item Number of Pathotypes
\item Simple Diversity Index
\item Gleason Diversity Index
\item Shannon Diversity Index
\item Simpson Diversity Index
\item Evenness Diversity Index
}
}
\description{
Calculates pathogen diversity indices.

This function calculates five diversity indices for the user.
\itemize{
\item Simple diversity index, which will show the proportion of unique pathotypes
to total samples. As the values gets closer to 1, there is greater diversity
in pathoypes within the population.
\deqn{S = \frac{No. pathotypes}{No. samples} }{S = No. pathotypes / No. samples }
\item Gleason diversity index, an alternate version of Simple diversity index.
This index is less sensitive to sample size than the simple index.
\deqn{G = \frac{ (No. pathotypes - 1) }{ log( No. samples)}}{G = (No. pathotypes -1) / log(No. Samples)}
\item Shannon diversity index, typically between 1.5 and 3.5. As richness and
\deqn{H' = -\sum_i p_i \log_{b} p_i}{ H = -sum p_i log(b) p_i }, where
\eqn{p_i} is the proportional abundance of species \eqn{i} and \eqn{b}
is the base of the logarithm.
\item Simpson diversity index values range from 0 to 1. 1 represents high
diversity and 0 represents no diversity.

\deqn{ D = \sum p_i^2 }{ D = sum p_i^2 }
\item Evenness ranges from 0 to 1. As the Eveness value approaches 1, there is a
more even distribution of each pathoypes frequency within the population.

\deqn{E = \frac{Shannon}{log(No. pathotypes) }}{E = Shannon / log(No. Pathotypes) }
}
}
\examples{
# Using the built-in data set, P_sojae_survey
data(P_sojae_survey)

P_sojae_survey

# calculate susceptibilities with a 60 \% cutoff value
diversities <- calculate_diversities(x = P_sojae_survey,
                                     cutoff = 60,
                                     control = "susceptible",
                                     sample = "Isolate",
                                     gene = "Rps",
                                     perc_susc = "perc.susc")

diversities

}
