% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{complexities}
\alias{complexities}
\title{Calculate distribution of complexities by \emph{Rps} gene}
\usage{
complexities(x, cutoff, control = "susceptible")
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Integer.}

\item{control}{Value used to denote the susceptible control in the \var{Rps}
field. Defaults to "susceptible". Character.}
}
\description{
This function will calculate the distribution of
susceptibilities by \emph{Rps} gene.
}
