% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{complexities}
\alias{complexities}
\title{Calculate distribution of complexities by \emph{Rps} gene}
\usage{
complexities(x, cutoff, control = "susceptible")
}
\arguments{
\item{x}{A \code{data.frame} containing the data. Character.}

\item{cutoff}{Value for percent susceptible cutoff. Integer.}

\item{control}{Value used to denote the susceptible control in the \var{Rps}
field. Defaults to "susceptible". Character.}
}
\description{
This function will calculate the distribution of
susceptibilities by \emph{Rps} gene.
}
\examples{


# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate totals and percents for resistant and susceptible reactions
Ps <- calc_percents(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calc_complexities(Ps, cutoff = 60)
complexities

}
