% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversities.R
\name{diversities_table}
\alias{diversities_table}
\title{Prints hagis.diversities Object Table of Diversities}
\usage{
diversities_table(x, type = "text", ...)
}
\arguments{
\item{x}{a hagis.diversities object generated by \code{\link[=calculate_diversities]{calculate_diversities()}}}

\item{type}{return a text or graphic type table. Valid values are \code{text} or
\code{graphic} Defaults to \code{text}.}

\item{...}{other arguments passed to \code{\link[gridExtra:ttheme_default]{gridExtra::ttheme_default()}}}
}
\description{
Print the frequency table of diversities from a \code{hagis.diversities} object.
Users may select from a text table (a \code{\link[pander:pander]{pander::pander()}} object) or as a
graphic, a \code{\link[gridExtra:tableGrob]{gridExtra::tableGrob()}} object
}
\examples{
# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
diversities <- calculate_diversities(x = Ps,
                                     cutoff = 60,
                                     control = "susceptible",
                                     sample = "Isolate",
                                     Rps = "Rps",
                                     perc_susc = "perc.susc")

# print the diversities table
diversities_table(diversities)

}
\seealso{
\code{\link[=calculate_diversities]{calculate_diversities()}}
}
