% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathotype_frequency_distribution.R
\name{freq_dist}
\alias{freq_dist}
\title{Calculate pathotype frequency distribution}
\usage{
freq_dist(x, sample, percent_susc, gene, susceptibility_cutoff)
}
\arguments{
\item{x}{A `data.frame`` containing the data}
}
\description{
Calculate pathotype frequency distribution
}
