% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hagis-package.R
\docType{package}
\name{hagis}
\alias{hagis}
\alias{hagis-package}
\title{Tools for Analysis of \emph{Phytophthora} Pathotype Complexities, Distributions and Diversity}
\description{
Provides tools for analysis of Phytophthora sojae survey data.  
  Functions provide distribution of susceptibilities, distribution of
  complexities with statistics, pathotype frequency distribution, as well as
  diversity indices for pathotypes.  This package is meant to be a direct
  replacement for the Habgood-Gilmour Spreadsheet, `HaGiS', previously used for
  `P. sojae' pathotype analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openplantpathology/hagis}
  \item Report bugs at \url{https://github.com/openplantpathology/hagis/issues}
}

}
\author{
Austin McCoy, mccoyaus@msu.edu and Zachary Noel, noelzach@msu.edu and
Adam H. Sparks, adam.sparks@usq.edu.au and Martin Chilvers, chilvers@msu.edu
}
