% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_sojae_survey}
\alias{P_sojae_survey}
\title{\emph{Phytophthora sojae} Survey Example Data}
\format{
An object of class \code{data.table} with 294 observations of 12 variables
\describe{
\item{Isolate}{\emph{P. sojae} isolate identifier}
\item{Line}{Soybean cultivar}
\item{Rps}{\emph{Rps} gene identifier}
\item{Total}{Total number of plants inoculated}
\item{HR (1)}{Number of plants that exhibit a hypersensitive response after
inoculation}
\item{Lesion (2)}{Number of plants that develop a lesion at inoculation
site}
\item{Lesion to cotyledon (3)}{Number of plants that develop a lesion,
which advances to the hypocotyl of the seedling after infection}
\item{Dead (4)}{Number of dead plants that are observed after inoculation}
\item{total.susc}{The total number of susceptible plants (Lesion+Lesion to
cotyledon+Dead)}
\item{total.resis}{The total number of resistant plants (equal to HR
value)}
}
}
\source{
Data from an ongoing 2017 \emph{Phytophthora sojae} survey in Michigan,
conducted by A. G. McCoy \emph{et al.}.
}
\usage{
data(P_sojae_survey)
}
\description{
Data from a \emph{Phytophthora sojae} survey
}
\examples{
data(P_sojae_survey)
P_sojae_survey

}
\keyword{datasets}
