% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{plot.hagis.complexities}
\alias{plot.hagis.complexities}
\title{Plot \pkg{hagis} Complexities}
\usage{
\method{plot}{hagis.complexities}(object, y)
}
\arguments{
\item{object}{a \pkg{hagis} \code{complexities} object generated by
\code{\link[=calculate_complexities]{calculate_complexities()}}. Character.}

\item{y}{a vector of values for which the bar plot is desired. Specify
whether to return a graph of the frequency of complexities as a percentage,
"\code{frequency}", or as the count, "\code{distribution}". Character.}
}
\value{
A \link[ggplot2]{ggplot2} plot
}
\description{
Creates a \pkg{ggplot2} object of the frequency of complexity
(percent per complexity) calculated by \code{\link[=calculate_complexities]{calculate_complexities()}} or a
\pkg{ggplot2} object of the distribution (number per complexity).
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate susceptibilities with a 60 \% cutoff value
complexities <- calculate_complexities(x = Ps,
                                      cutoff = 60,
                                      control = "susceptible",
                                      sample = "Isolate",
                                      Rps = "Rps",
                                      perc_susc = "perc.susc")

# Visualize the distribution (actual values)
plot(object = complexities, y = "distribution")

# Visualize the frequency (percentages)
plot(object = complexities, y = "frequency")

}
\seealso{
\code{\link[=calculate_complexities]{calculate_complexities()}}
}
