% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexities.R
\name{print.summary.hagis}
\alias{print.summary.hagis}
\title{Prints summary.hagis Object for Complexities}
\usage{
\method{print}{summary.hagis}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{a summary.hagis object}

\item{...}{ignored}
}
\description{
Custom \code{\link[=print]{print()}} method for \pkg{hagis} \code{complexity} objects.
}
