% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rps.R
\name{summarize_rps}
\alias{summarize_rps}
\title{Calculate and Summarize Distribution of Susceptibilities by \emph{Rps} Gene}
\usage{
summarize_rps(x, cutoff, control, sample, Rps, perc_susc)
}
\arguments{
\item{x}{a \code{data.frame} containing the data.}

\item{cutoff}{value for percent susceptible cutoff. Numeric.}

\item{control}{value used to denote the susceptible control in the \code{Rps}
field. Character.}

\item{sample}{field providing the unique isolate identification for each
isolate being tested. Character.}

\item{Rps}{field providing the \emph{Rps} gene(s) being tested. Character.}

\item{perc_susc}{field providing the percent susceptible reactions.
Character.}
}
\value{
returns an object of \code{\link[=class]{class()}} \code{hagis.rps.summary}
An object of class \code{hagis.summaries} is a \link[=data.table]{data.table}
containing the following components fields
\describe{
\item{Rps}{the Rps gene}
\item{N_susc}{the total number susceptible for a given \emph{Rps} gene in the
\code{Rps} field}
\item{percent_pathogenic}{the frequency with which an \emph{Rps} gene is
pathogenic}
}
}
\description{
This function will calculate the distribution of
susceptibilities by \emph{Rps} gene.
}
\examples{

# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_rps(x = Ps,
                       cutoff = 60,
                       control = "susceptible",
                       sample = "Isolate",
                       Rps = "Rps",
                       perc_susc = "perc.susc")
susc

# plot susceptibilities
autoplot(susc)

}
