% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_susceptiblities.R
\name{summarize_susc}
\alias{summarize_susc}
\title{Calculate and Summarize Distribution of Susceptibilities by Pathogenicity Gene}
\usage{
summarize_susc(x, cutoff, control = "susceptible", sample = "sample",
  gene = "gene", perc_susc = "perc.susc")
}
\arguments{
\item{x}{A \code{data.frame} containing the data. See Details below for more.
Character.}

\item{cutoff}{Value for percent susceptible cutoff. Numeric.}

\item{control}{Value used to denote the susceptible control in the \code{gene}
field. Defaults to "susceptible". Character.}

\item{sample}{Field providing the unique isolate identification for each
isolate being tested. Defaults to \code{sample}. Character.}

\item{gene}{Field providing the pathogenicty gene(s) being tested. Defaults
to \code{gene}. Character.}

\item{perc_susc}{Field providing the percent susceptible reactions. Defaults
to \code{perc.susc}. Character.

The use of \code{\link[=write_template]{write_template()}} and \code{\link[=calc_percents]{calc_percents()}} is encouraged to ensure
that the input \code{data.frame}, \var{x}, used is properly formatted.}
}
\description{
This function will calculate the distribution of
susceptibilities by pathogenicity gene.
}
\examples{

# locate system file for import
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)
head(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_susc(Ps, cutoff = 60, sample = "Isolate", gene = "Rps")
susc

}
\seealso{
\code{\link[=write_template]{write_template()}}, \code{\link[=percents]{percents()}}, \code{\link[=visualize_susc]{visualize_susc()}}
}
