% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_complexities.R
\name{summary.hagis}
\alias{summary.hagis}
\title{Summarises hagis complexity object.}
\usage{
\method{summary}{hagis}(object, ...)
}
\arguments{
\item{...}{ignored}

\item{x}{A hagis class object (a list of two \code{data.table}s)}
}
\description{
Summarises hagis complexity object.
}
