% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_of_susceptibilities.R
\name{susceptibilities}
\alias{susceptibilities}
\title{Calculate distribution of susceptibilities by Rps gene}
\usage{
susceptibilities(x, sample, perc.susc, gene, susc_cutoff)
}
\arguments{
\item{x}{A \code{data.frame} containing the data}

\item{susc_cutoff}{}
}
\description{
This function will calculate the distribution of
susceptibilities by Rps gene.
}
