% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_susceptiblities.R
\name{visualize_Rps}
\alias{visualize_Rps}
\title{Visualize \emph{Rps} Gene Summaries}
\usage{
visualize_Rps(x)
}
\arguments{
\item{x}{A \code{data.frame} generated by \code{\link[=summarize_Rps]{summarize_Rps()}}. Character.}
}
\description{
Creates a \pkg{ggplot2} object of the \emph{Rps} gene summaries
calculated by \code{\link[=summarize_Rps]{summarize_Rps()}}
}
\details{
The \code{data.frame} object supplied should be generated by \code{\link[=summarize_Rps]{summarize_Rps()}} and
contain the following fields.
\itemize{
\item Rps - \emph{Rps} gene(s) included in study
\item N - Total number of plants susceptible
\item percent.pathogenic - Percent of isolates pathogenic on \emph{Rps} genes
}
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate totals and percents for resistant and susceptible reactions
Ps <- calc_percents(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_Rps(Ps, cutoff = 60)

# Visualize the summary of _Rps_ genes
visualize_Rps(susc)

}
\seealso{
\code{\link[=write_template]{write_template()}}, \code{\link[=calc_percents]{calc_percents()}}, \code{\link[=summarize_Rps]{summarize_Rps()}}
}
