% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_susceptiblities.R
\name{visualize_susc}
\alias{visualize_susc}
\title{Visualize Pathogenicity Gene Summaries}
\usage{
visualize_susc(x)
}
\arguments{
\item{x}{A \code{data.frame} generated by \code{\link[=summarize_susc]{summarize_susc()}}. Character.}
}
\description{
Creates a \pkg{ggplot2} object of the pathogenicity gene
summaries calculated by \code{\link[=summarize_susc]{summarize_susc()}}
}
\details{
The \code{data.frame} object supplied should be generated by \code{\link[=summarize_susc]{summarize_susc()}} and
contain the following fields.
\itemize{
\item gene - Pathogenicity genes included in study
\item N - Total number of plants susceptible
\item percent.pathogenic - Percent of isolates pathogenic on pathogenicity genes
}
}
\examples{
Ps <- system.file("extdata", "practice_data_set.csv", package = "hagis")

# import 'practice_data_set.csv'
Ps <- read.csv(Ps)

# calculate totals and percents for resistant and susceptible reactions
Ps <- calc_percents(Ps)

# calculate susceptibilities with a 60 \% cutoff value
susc <- summarize_susc(Ps, cutoff = 60)

# Visualize the summary of pathogenicity genes
visualize_susc(susc)

}
\seealso{
\code{\link[=write_template]{write_template()}}, \code{\link[=calc_percents]{calc_percents()}}, \code{\link[=summarize_susc]{summarize_susc()}}
}
