% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{write_template}
\alias{write_template}
\title{Create a Phytophtora Template to Enter Data for Use with hagis
Creates a template  'CSV' file for the user to enter observation data into to
use for analysis This helps ensure that data are properly formatted for use
with \pkg{hagis}.}
\usage{
write_template(file = NULL)
}
\arguments{
\item{file}{A file path and filename for the template, defaults to
"Phytopthora_template.csv" in your current working directory.}
}
\description{
Create a Phytophtora Template to Enter Data for Use with hagis
Creates a template  'CSV' file for the user to enter observation data into to
use for analysis This helps ensure that data are properly formatted for use
with \pkg{hagis}.
}
\examples{
\donttest{
# Create a template in user's home subdirectory called "P_sojae_analysis"
write_template("~/P_sojae_analysis")
}

}
